test(NODE-3777): move csfle to manual
