+++ test/manual/client-side-encryption/client_side_encryption.corpus.spec.test.js
