+++ test/manual/client-side-encryption/client_side_encryption.prose.deadlock.js
@@ -2,7 +2,7 @@
 
 const BSON = require('bson');
 const { expect } = require('chai');
-const { dropCollection } = require('../shared');
+const { dropCollection } = require('../../integration/shared');
 const util = require('util');
 const fs = require('fs');
 const path = require('path');
