+++ test/manual/client-side-encryption/client_side_encryption.spec.test.js
